Gave searchBar its own component
